+++ include/JoSIM/Resistor.hpp
@@ -45,7 +45,7 @@ namespace JoSIM {
     Resistor(
       const std::pair<tokens_t, string_o>& s, const NodeConfig& ncon,
       const nodemap& nm, std::unordered_set<std::string>& lm, 
-      nodeconnections& nc, Input& iObj, Spread& spread, int& bi);
+      nodeconnections& nc, Input& iObj, Spread& spread, int64_t& bi);
 
     void update_timestep(const double& factor) override;
